+++ examples/GeneratorFormats.cpp
@@ -37,6 +37,9 @@ int main()
     // Generate a scramble.
     std::cout << gen.scramble(24, 'a', 'z') << '\n';
 
+    // Generate an IPv6 IP address with a port.
+    std::cout << gen.ip_addr(lpsm::IPV6, lpsm::PORT) << '\n';
+
     // Generate 5 Markdown paragraphs.
     std::cout << gen.fmt_paragraph(5, lpsm::USELIPSUM, lpsm::MARKDOWN);